Fix a warning that sneaked in
Related to MIS stuff I hastily merged, that surface area value isn't
even used for anything yet.

+++ src/utils/loaders/formats/wavefront/wavefront.c
@@ -257,7 +257,7 @@ struct mesh *parseWavefront(const char *filePath, size_t *finalMeshCount, struct
 		}
 	}
 
-	logr(debug, "Mesh %s surface area is %.4fm²\n", currentMeshPtr->name, surface_area);
+	logr(debug, "Mesh %s surface area is %.4fm²\n", currentMeshPtr->name, (double)surface_area);
 	
 	currentMeshPtr->surface_area = surface_area;
 	currentMeshPtr->polygons = polygons;
